Add ternary and null coalescing operator support
The module contains the Invoke-Assignment function created by Garret Serack
(@FearTheCowboy) [1] It uses powershell aliases combined with eval.

The alias '==' might not be exported properly by the script module.
After importing the invoke-assignment module it can be added by
    Set-Alias = Invoke-Assignment
Maybe an alias like ~= is preferable, so that the resulting assignment
operator is '=~=' instead of '=='. The latter could be confused with an
equals operator.

[1] http://fearthecowboy.com/powershell/2015/12/11/PowerShell-A-Better-Ternary/