Default to calling script's dir instead of workdir
The New-Package cmdlet used the current working directory as default
value for several parameters. In case the cmdlet is invoked by a
script, this has been changed to the directory containing the calling
script.

This is especially useful for the TemplateDir-Parameter: From now on it
is no longer necessary to set this parameter when using New-Package in
template build scripts.